Fold CodeRewriter into RewriteCsFilesStep
CodeRewriter.Rewrite had a single caller (RewriteCsFilesStep). Move its regex
fields and Rewrite method directly into the step and delete the now-empty class.
Rename the spec file/class to match.

Mirrors the CsprojRewriter fold, so a migration step owns the rewrite rules for
its artifact type instead of delegating them to a free-floating static class in
Steps/. Raised in review on #528.

No behaviour change.

(cherry picked from commit 9fb8800529cc39037d9ad0febae8475d7d7a9c36)
